refactor(opencode-plugin): separate server tui and tests
