--- opencode-plugin/src/.npmignore
@@ -1,1 +1,0 @@
-*.test.ts
