+++ opencode-plugin/src/server/debug.ts
