+++ opencode-plugin/src/server/hooks.ts
