+++ opencode-plugin/src/server/ingest/select.ts
