+++ opencode-plugin/src/server/ingest/submit.ts
