+++ opencode-plugin/src/server/recall/query.ts
