+++ opencode-plugin/src/server/server-backend.ts
@@ -10,7 +10,7 @@ import type {
   CreateMemoryInput,
   UpdateMemoryInput,
   SearchInput,
-} from "./types.js";
+} from "../shared/types.js";
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
