+++ opencode-plugin/src/server/session-transcript.ts
