+++ opencode-plugin/src/server/setup-flow.ts
