+++ opencode-plugin/src/server/tools.ts
@@ -1,6 +1,10 @@
 import { tool } from "@opencode-ai/plugin";
 import type { MemoryBackend } from "./backend.js";
-import type { CreateMemoryInput, UpdateMemoryInput, SearchInput } from "./types.js";
+import type {
+  CreateMemoryInput,
+  UpdateMemoryInput,
+  SearchInput,
+} from "../shared/types.js";
 
 /**
  * Build the 5 memory tools for OpenCode.
