+++ opencode-plugin/src/shared/credentials-store.ts
