+++ opencode-plugin/src/shared/defaults.ts
