+++ opencode-plugin/src/shared/platform-paths.ts
