+++ opencode-plugin/src/shared/plugin-meta.ts
