+++ opencode-plugin/src/shared/plugin-spec.ts
