+++ opencode-plugin/src/shared/setup-files.ts
