+++ opencode-plugin/src/shared/types.ts
